IGNITE-18349 Move LogicalTopologyService to an internal package
