+++ modules/cluster-management/src/main/java/org/apache/ignite/internal/cluster/management/topology/api/LogicalTopologyService.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.network;
+package org.apache.ignite.internal.cluster.management.topology.api;
 
 import java.util.concurrent.CompletableFuture;
+import org.apache.ignite.network.TopologyService;
 
 /**
  * Used for getting information about the cluster's Logical Topology.
